requirements.txt
aiogram==3.0.0b7
openai==0.27.0
python-dotenv==0.20.0
APScheduler==3.9.1.post1
